move pq_rl_helper
