+++ ydb/core/persqueue/public/ya.make
@@ -2,6 +2,7 @@ LIBRARY()
 
 SRCS(
     pq_database.cpp
+    pq_rl_helpers.cpp
     utils.cpp
 )
