+++ ydb/core/persqueue/writer/writer.h
@@ -5,7 +5,7 @@
 #include <ydb/core/grpc_services/local_rate_limiter.h>
 #include <ydb/core/protos/msgbus.pb.h>
 #include <ydb/core/protos/msgbus_pq.pb.h>
-#include <ydb/core/persqueue/pq_rl_helpers.h>
+#include <ydb/core/persqueue/public/pq_rl_helpers.h>
 #include <ydb/core/persqueue/write_id.h>
 #include <ydb/core/kafka_proxy/kafka_producer_instance_id.h>
