+++ ydb/services/datastreams/datastreams_proxy.cpp
@@ -9,7 +9,7 @@
 #include <ydb/core/grpc_services/rpc_deferrable.h>
 #include <ydb/core/grpc_services/rpc_scheme_base.h>
 #include <ydb/core/persqueue/pqtablet/partition/partition.h>
-#include <ydb/core/persqueue/pq_rl_helpers.h>
+#include <ydb/core/persqueue/public/pq_rl_helpers.h>
 #include <ydb/core/persqueue/write_meta.h>
 #include <ydb/core/persqueue/events/internal.h>
