+++ ydb/services/persqueue_v1/actors/direct_read_actor.h
@@ -7,7 +7,7 @@
 #include <ydb/core/grpc_services/grpc_request_proxy.h>
 #include <ydb/core/persqueue/events/global.h>
 
-#include <ydb/core/persqueue/pq_rl_helpers.h>
+#include <ydb/core/persqueue/public/pq_rl_helpers.h>
 
 #include <ydb/library/actors/core/actor_bootstrapped.h>
